# src/ApiDocs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v0.0.5",
    "title": "Shopping backend",
    "description": "Backend for a web shop"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/dbtest": {
      "get": {
        "tags": [
          "Test"
        ],
        "summary": "Test db authentication",
        "description": "Shows if the db is authentication",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/test/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/testMail/": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/products/list": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List cards",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/products/search/{query}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Search cards from string",
        "description": "",
        "parameters": [
          {
            "name": "query",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get card details",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/basket/{sessionId?}": {
      "get": {
        "tags": [
          "Basket"
        ],
        "summary": "Get basket",
        "description": "",
        "parameters": [
          {
            "name": "sessionId?",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/basket/add": {
      "post": {
        "tags": [
          "Basket"
        ],
        "summary": "Add item to basket",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "sessionId": {
                  "example": "any"
                },
                "itemId": {
                  "example": "any"
                },
                "quantity": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/basket/{sessionId}/{item_id}": {
      "delete": {
        "tags": [
          "Basket"
        ],
        "summary": "Remove item from basket",
        "description": "",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/basket/{sessionId}/": {
      "delete": {
        "tags": [
          "Basket"
        ],
        "summary": "Delete basket",
        "description": "",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/basket/order": {
      "post": {
        "tags": [
          "Basket"
        ],
        "summary": "Place order",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "sessionId": {
                  "example": "any"
                },
                "paymentMethod": {
                  "example": "any"
                },
                "streetName": {
                  "example": "any"
                },
                "city": {
                  "example": "any"
                },
                "zipCode": {
                  "example": "any"
                },
                "fullName": {
                  "example": "any"
                },
                "phoneNumber": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/basket/order/receipt/{ordernumber}": {
      "get": {
        "tags": [
          "Basket"
        ],
        "summary": "Get receipt basket",
        "description": "Maybe this should be post idk?",
        "parameters": [
          {
            "name": "ordernumber",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}